requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv5/YOLOv8
timm>=0.9.0  # ResNet pretrained models

# Data processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utils
tqdm>=4.66.0
PyYAML>=6.0